ai/requirements.txt
salesforce-lavis
deepspeech
torch
google-generativeai
pydub
protoc_gen_openapiv2
